Remove getParameterByName global
