+++ src/components/favoriteitems.js
@@ -5,6 +5,7 @@ import { appHost } from './apphost';
 import imageLoader from './images/imageLoader';
 import globalize from '../scripts/globalize';
 import layoutManager from './layoutManager';
+import { getParameterByName } from '../utils/url.ts';
 import '../assets/css/scrollstyles.scss';
 import '../elements/emby-itemscontainer/emby-itemscontainer';
