+++ src/controllers/dashboard/dlna/profile.js
@@ -9,6 +9,7 @@ import '../../../elements/emby-checkbox/emby-checkbox';
 import '../../../components/listview/listview.scss';
 import Dashboard from '../../../scripts/clientUtils';
 import toast from '../../../components/toast/toast';
+import { getParameterByName } from '../../../utils/url.ts';
 
 /* eslint-disable indent */
