+++ src/controllers/dashboard/notifications/notification/index.js
@@ -2,6 +2,7 @@ import escapeHtml from 'escape-html';
 import 'jquery';
 import '../../../../elements/emby-checkbox/emby-checkbox';
 import Dashboard from '../../../../scripts/clientUtils';
+import { getParameterByName } from '../../../../utils/url.ts';
 
 function fillItems(elem, items, cssClass, idPrefix, currentList, isEnabledList) {
     let html = '<div class="checkboxList paperList" style="padding: .5em 1em;">';
