+++ src/controllers/dashboard/scheduledtasks/scheduledtask.js
@@ -7,6 +7,7 @@ import '../../../elements/emby-input/emby-input';
 import '../../../elements/emby-button/emby-button';
 import '../../../elements/emby-select/emby-select';
 import confirm from '../../../components/confirm/confirm';
+import { getParameterByName } from '../../../utils/url.ts';
 
 /* eslint-disable indent */
