+++ src/controllers/livetvguideprovider.js
@@ -2,6 +2,7 @@ import { Events } from 'jellyfin-apiclient';
 import loading from '../components/loading/loading';
 import globalize from '../scripts/globalize';
 import Dashboard, { pageIdOn } from '../scripts/clientUtils';
+import { getParameterByName } from '../utils/url.ts';
 
 function onListingsSubmitted() {
     Dashboard.navigate('livetvstatus.html');
