+++ src/controllers/livetvtuner.js
@@ -6,6 +6,7 @@ import '../elements/emby-button/emby-button';
 import '../elements/emby-checkbox/emby-checkbox';
 import '../elements/emby-select/emby-select';
 import Dashboard from '../scripts/clientUtils';
+import { getParameterByName } from '../utils/url.ts';
 
 function isM3uVariant(type) {
     return ['nextpvr'].indexOf(type || '') !== -1;
