+++ src/scripts/libraryMenu.js
@@ -19,6 +19,7 @@ import '../assets/css/flexstyles.scss';
 import Dashboard, { pageClassOn } from './clientUtils';
 import ServerConnections from '../components/ServerConnections';
 import Headroom from 'headroom.js';
+import { getParameterByName } from '../utils/url.ts';
 
 /* eslint-disable indent */
