+++ src/utils/url.ts
@@ -1,4 +1,4 @@
-export const getWindowLocationSearch = (win: Window | null | undefined) => {
+export const getWindowLocationSearch = (win?: Window | null | undefined) => {
     let search = (win || window).location.search;
 
     if (!search) {
@@ -10,4 +10,17 @@ export const getWindowLocationSearch = (win: Window | null | undefined) => {
     }
 
     return search || '';
+};
+
+export const getParameterByName = (name: string, url?: string | null | undefined) => {
+    name = name.replace(/[[]/, '\\[').replace(/[\]]/, '\\]');
+    const regexS = '[\\?&]' + name + '=([^&#]*)';
+    const regex = new RegExp(regexS, 'i');
+    const results = regex.exec(url || getWindowLocationSearch());
+
+    if (results == null) {
+        return '';
+    }
+
+    return decodeURIComponent(results[1].replace(/\+/g, ' '));
 };
